# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Character {
  id: number;
  name: string;
  className: string;
  race: string;
  level: number;
  maxHp: number;
  currentHp: number;
  armorClass: number;
  speed: number;
  strength: number;
  dexterity: number;
  constitution: number;
  intelligence: number;
  wisdom: number;
  charisma: number;
  /** @nullable */
  background?: string | null;
  /** @nullable */
  backstory?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CharacterInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  className: string;
  /** @minLength 1 */
  race: string;
  /** @minimum 1 */
  level: number;
  /** @minimum 1 */
  maxHp: number;
  currentHp: number;
  armorClass: number;
  speed: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  strength: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  dexterity: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  constitution: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  intelligence: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  wisdom: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  charisma: number;
  background?: string;
  backstory?: string;
}

export interface CharacterUpdate {
  /** @minLength 1 */
  name?: string;
  className?: string;
  race?: string;
  /** @minimum 1 */
  level?: number;
  /** @minimum 1 */
  maxHp?: number;
  currentHp?: number;
  armorClass?: number;
  speed?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  strength?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  dexterity?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  constitution?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  intelligence?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  wisdom?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  charisma?: number;
  background?: string;
  backstory?: string;
}

export type RollDiceType = typeof RollDiceType[keyof typeof RollDiceType];


export const RollDiceType = {
  d4: 'd4',
  d6: 'd6',
  d8: 'd8',
  d10: 'd10',
  d12: 'd12',
  d20: 'd20',
  d100: 'd100',
} as const;

export interface Roll {
  id: number;
  characterId: number;
  diceType: RollDiceType;
  result: number;
  /** @nullable */
  modifier?: number | null;
  total?: number;
  /** @nullable */
  label?: string | null;
  rolledAt: string;
}

export type RollInputDiceType = typeof RollInputDiceType[keyof typeof RollInputDiceType];


export const RollInputDiceType = {
  d4: 'd4',
  d6: 'd6',
  d8: 'd8',
  d10: 'd10',
  d12: 'd12',
  d20: 'd20',
  d100: 'd100',
} as const;

export interface RollInput {
  diceType: RollInputDiceType;
  modifier?: number;
  label?: string;
}

export type RollWithCharacterDiceType = typeof RollWithCharacterDiceType[keyof typeof RollWithCharacterDiceType];


export const RollWithCharacterDiceType = {
  d4: 'd4',
  d6: 'd6',
  d8: 'd8',
  d10: 'd10',
  d12: 'd12',
  d20: 'd20',
  d100: 'd100',
} as const;

export interface RollWithCharacter {
  id: number;
  characterId: number;
  characterName: string;
  diceType: RollWithCharacterDiceType;
  result: number;
  /** @nullable */
  modifier?: number | null;
  total: number;
  /** @nullable */
  label?: string | null;
  rolledAt: string;
}

export type NoteCategory = typeof NoteCategory[keyof typeof NoteCategory];


export const NoteCategory = {
  location: 'location',
  npc: 'npc',
  item: 'item',
  lore: 'lore',
  general: 'general',
} as const;

export interface Note {
  id: number;
  title: string;
  content: string;
  category: NoteCategory;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}

export type NoteInputCategory = typeof NoteInputCategory[keyof typeof NoteInputCategory];


export const NoteInputCategory = {
  location: 'location',
  npc: 'npc',
  item: 'item',
  lore: 'lore',
  general: 'general',
} as const;

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  content: string;
  category: NoteInputCategory;
  tags?: string[];
}

export type NoteUpdateCategory = typeof NoteUpdateCategory[keyof typeof NoteUpdateCategory];


export const NoteUpdateCategory = {
  location: 'location',
  npc: 'npc',
  item: 'item',
  lore: 'lore',
  general: 'general',
} as const;

export interface NoteUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  category?: NoteUpdateCategory;
  tags?: string[];
}