# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface Character {
  id: number;
  name: string;
  className: string;
  race: string;
  level: number;
  maxHp: number;
  currentHp: number;
  armorClass: number;
  speed: number;
  strength: number;
  dexterity: number;
  constitution: number;
  intelligence: number;
  wisdom: number;
  charisma: number;
  /** @nullable */
  background?: string | null;
  /** @nullable */
  backstory?: string | null;
  createdAt: Date;
  updatedAt: Date;
}