# lib/api-zod/src/generated/types/characterUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface CharacterUpdate {
  /** @minLength 1 */
  name?: string;
  className?: string;
  race?: string;
  /** @minimum 1 */
  level?: number;
  /** @minimum 1 */
  maxHp?: number;
  currentHp?: number;
  armorClass?: number;
  speed?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  strength?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  dexterity?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  constitution?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  intelligence?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  wisdom?: number;
  /**
     * @minimum 1
     * @maximum 30
     */
  charisma?: number;
  background?: string;
  backstory?: string;
}