# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { NoteCategory } from './noteCategory';

export interface Note {
  id: number;
  title: string;
  content: string;
  category: NoteCategory;
  tags?: string[];
  createdAt: Date;
  updatedAt: Date;
}