# lib/api-zod/src/generated/types/noteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { NoteInputCategory } from './noteInputCategory';

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  content: string;
  category: NoteInputCategory;
  tags?: string[];
}