# lib/api-zod/src/generated/types/noteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { NoteUpdateCategory } from './noteUpdateCategory';

export interface NoteUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  category?: NoteUpdateCategory;
  tags?: string[];
}