# lib/api-zod/src/generated/types/roll.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { RollDiceType } from './rollDiceType';

export interface Roll {
  id: number;
  characterId: number;
  diceType: RollDiceType;
  result: number;
  /** @nullable */
  modifier?: number | null;
  total?: number;
  /** @nullable */
  label?: string | null;
  rolledAt: Date;
}