# lib/api-zod/src/generated/types/rollInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { RollInputDiceType } from './rollInputDiceType';

export interface RollInput {
  diceType: RollInputDiceType;
  modifier?: number;
  label?: string;
}