# lib/api-zod/src/generated/types/rollWithCharacter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPG Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { RollWithCharacterDiceType } from './rollWithCharacterDiceType';

export interface RollWithCharacter {
  id: number;
  characterId: number;
  characterName: string;
  diceType: RollWithCharacterDiceType;
  result: number;
  /** @nullable */
  modifier?: number | null;
  total: number;
  /** @nullable */
  label?: string | null;
  rolledAt: Date;
}